@@ -15,9 +15,7 @@
     "icon": "ng g ng-alain:plugin icon",
     "lint:ts": "ng lint --fix",
     "lint:style": "stylelint \"src/**/*.less\" --syntax less --fix",
-    "lint-staged": "lint-staged",
-    "tslint-check": "tslint-config-prettier-check ./tslint.json",
-    "hmr": "ng serve -c=hmr"
+    "tslint-check": "tslint-config-prettier-check ./tslint.json"
   },
   "private": true,
   "dependencies": {
@@ -30,26 +28,26 @@
     "@angular/platform-browser-dynamic": "11.2.5",
     "@angular/platform-server": "11.2.5",
     "@angular/router": "11.2.5",
-    "@ant-design/icons-angular": "^11.0.1",
-    "@delon/abc": "^9.5.5",
-    "@delon/acl": "^9.5.5",
-    "@delon/auth": "^9.5.5",
-    "@delon/cache": "^9.5.5",
-    "@delon/chart": "^9.5.5",
-    "@delon/form": "^9.5.5",
-    "@delon/mock": "^9.5.5",
-    "@delon/theme": "^9.5.5",
-    "@delon/util": "^9.5.5",
-    "@ngx-translate/core": "^12.1.2",
-    "ajv": "^6.12.0",
+    "@delon/abc": "^11.8.0",
+    "@delon/acl": "^11.8.0",
+    "@delon/auth": "^11.8.0",
+    "@delon/cache": "^11.8.0",
+    "@delon/chart": "^11.8.0",
+    "@delon/form": "^11.8.0",
+    "@delon/mock": "^11.8.0",
+    "@delon/theme": "^11.8.0",
+    "@delon/util": "^11.8.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
+    "ajv": "^6.12.6",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.4.1",
     "core-js": "^3.3.3",
     "date-fns": "^2.12.0",
     "jquery": "3.5.0",
     "ng-zorro-antd": "^11.2.0",
-    "ngx-tinymce": "^9.0.0",
-    "ngx-ueditor": "^9.0.0",
+    "ngx-tinymce": "^11.0.0",
+    "ngx-ueditor": "^11.0.0",
     "oidc-client": "^1.9.1",
     "popper.js": "^1.16.0",
     "rxjs": "^6.5.3",
@@ -58,12 +56,11 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.4",
+    "@angular-devkit/build-angular": "~0.1102.3",
     "@angular/cli": "^11.2.4",
     "@angular/compiler-cli": "11.2.5",
     "@angular/language-service": "11.2.5",
-    "@angularclass/hmr": "^2.1.3",
-    "@delon/testing": "^8.8.0",
+    "@delon/testing": "^11.8.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/jszip": "^3.1.7",
@@ -80,7 +77,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "ng-alain": "^11.8.0",
     "ng-alain-codelyzer": "^0.0.1",
-    "typescript": "^4.0.7",
+    "typescript": "~4.1.5",
     "xlsx": "^0.15.6"
   },
   "optionalDependencies": {
@@ -90,7 +87,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit-remove": "npm run lint-staged"
+      "pre-commit-remove": ""
     }
   }
 }
