@@ -1,14 +1,14 @@
 // tslint:disable: no-duplicate-imports
-import { NgModule, LOCALE_ID, APP_INITIALIZER  } from '@angular/core';
-import { HttpClientModule  } from '@angular/common/http';
+import { NgModule, LOCALE_ID, APP_INITIALIZER } from '@angular/core';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 // #region default language
 // Reference: https://ng-alain.com/docs/i18n
 import { default as ngLang } from '@angular/common/locales/zh';
-import { NZ_I18N, zh_CN as zorroLang } from 'ng-zorro-antd/i18n';
-import { DELON_LOCALE, zh_CN as delonLang } from '@delon/theme';
+import { NZ_I18N, en_US as zorroLang } from 'ng-zorro-antd/i18n';
+import { DELON_LOCALE, en_US as delonLang } from '@delon/theme';
 
 const LANG = {
   abbr: 'zh',
@@ -17,14 +17,39 @@ const LANG = {
   delon: delonLang,
 };
 // register angular
-//import { registerLocaleData } from '@angular/common';
-//registerLocaleData(LANG.ng, LANG.abbr);
+import { registerLocaleData } from '@angular/common';
+registerLocaleData(LANG.ng, LANG.abbr);
 
 const LANG_PROVIDES = [
   {provide: LOCALE_ID, useValue: LANG.abbr},
   {provide: NZ_I18N, useValue: LANG.zorro},
   {provide: DELON_LOCALE, useValue: LANG.delon}
 ];
+// #endregion
+
+// #region i18n services
+import { I18NService } from '@core';
+import { ALAIN_I18N_TOKEN } from '@delon/theme';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+
+// 加载i18n语言文件
+export function I18nHttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
+  return new TranslateHttpLoader(http, `assets/i18n/`, '.json');
+}
+
+const I18NSERVICE_MODULES = [
+  TranslateModule.forRoot({
+    loader: {
+      provide: TranslateLoader,
+      useFactory: I18nHttpLoaderFactory,
+      deps: [HttpClient],
+    },
+  }),
+];
+
+const I18NSERVICE_PROVIDES = [{ provide: ALAIN_I18N_TOKEN, useClass: I18NService, multi: false }];
+
 // #endregion
 
 // #region JSON Schema form (using @delon/form)
@@ -36,8 +61,15 @@ const FORM_MODULES = [JsonSchemaModule];
 
 // #region Http Interceptors
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import {  DefaultInterceptor } from '@core';
+import { DefaultInterceptor } from '@core';
 
+import { CoreModule } from '@core/core.module';
+import { SharedModule } from '@shared';
+import { AppComponent } from './app.component';
+import { RoutesModule } from './routes/routes.module';
+import { LayoutModule } from './layout/layout.module';
+import { GlobalConfigModule } from "./global-config.module";
+import { JWTInterceptor } from "@delon/auth";
 
 const INTERCEPTOR_PROVIDES = [
   {provide: HTTP_INTERCEPTORS, useClass: JWTInterceptor, multi: true},
@@ -65,15 +97,9 @@ const APPINIT_PROVIDES = [
     multi: true
   }
 ];
+
 // #endregion
 
-import { CoreModule } from '@core/core.module';
-import { SharedModule } from '@shared';
-import { AppComponent } from './app.component';
-import { RoutesModule } from './routes/routes.module';
-import { LayoutModule } from './layout/layout.module';
-import { JWTInterceptor } from "@delon/auth";
-import { GlobalConfigModule } from "./global-config.module";
 
 @NgModule({
   declarations: [
@@ -88,12 +114,14 @@ import { GlobalConfigModule } from "./global-config.module";
     SharedModule,
     LayoutModule,
     RoutesModule,
+    I18NSERVICE_MODULES,
     ...FORM_MODULES,
     ...GLOBAL_THIRD_MODULES
   ],
   providers: [
     ...LANG_PROVIDES,
     ...INTERCEPTOR_PROVIDES,
+    ...I18NSERVICE_PROVIDES,
     ...APPINIT_PROVIDES,
   ],
   bootstrap: [AppComponent]
