@@ -1,4 +1,4 @@
-
+export * from './i18n/i18n.service';
 export * from './module-import-guard';
 export * from './net/default.interceptor';
 export * from './startup/startup.service';
