@@ -1,11 +1,11 @@
-import { Component, HostListener, ChangeDetectionStrategy } from '@angular/core';
+import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
 import * as screenfull from 'screenfull';
 
 @Component({
   selector: 'header-fullscreen',
   template: `
     <i nz-icon [nzType]="status ? 'fullscreen-exit' : 'fullscreen'"></i>
-    {{ status ? '退出全屏' : '全屏' }}
+    {{ (status ? 'menu.fullscreen.exit' : 'menu.fullscreen') | translate }}
   `,
   // tslint:disable-next-line: no-host-metadata-property
   host: {
@@ -20,12 +20,12 @@ export class HeaderFullScreenComponent {
   }
 
   @HostListener('window:resize')
-  _resize() {
+  _resize(): void {
     this.status = this.sf.isFullscreen;
   }
 
   @HostListener('click')
-  _click() {
+  _click(): void {
     if (this.sf.isEnabled) {
       this.sf.toggle();
     }
