@@ -1,10 +1,11 @@
-import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
-import { NoticeItem, NoticeIconList } from '@delon/abc';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from '@angular/core';
+import { NoticeIconList, NoticeIconSelect, NoticeItem } from '@delon/abc/notice-icon';
+import add from 'date-fns/add';
+import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import parse from 'date-fns/parse';
+import { NzI18nService } from 'ng-zorro-antd/i18n';
+import { NzMessageService } from 'ng-zorro-antd/message';
 
-/**
- * 菜单通知
- */
 @Component({
   selector: 'header-notify',
   template: `
@@ -48,51 +49,60 @@ export class HeaderNotifyComponent {
   count = 5;
   loading = false;
 
-  constructor(private msg: NzMessageService, private cdr: ChangeDetectorRef) {}
+  constructor(private msg: NzMessageService, private nzI18n: NzI18nService, private cdr: ChangeDetectorRef) {}
 
   private updateNoticeData(notices: NoticeIconList[]): NoticeItem[] {
     const data = this.data.slice();
-    data.forEach(i => (i.list = []));
+    data.forEach((i) => (i.list = []));
 
-    notices.forEach(item => {
-      const newItem = { ...item };
+    notices.forEach((item) => {
+      const newItem = { ...item } as NoticeIconList;
+      if (typeof newItem.datetime === 'string') {
+        newItem.datetime = parse(newItem.datetime, 'yyyy-MM-dd', new Date());
+      }
+      if (newItem.datetime) {
+        newItem.datetime = formatDistanceToNow(newItem.datetime as Date, { locale: this.nzI18n.getDateLocale() });
+      }
       if (newItem.extra && newItem.status) {
-        newItem.color = {
+        newItem.color = ({
           todo: undefined,
           processing: 'blue',
           urgent: 'red',
           doing: 'gold',
-        }[newItem.status];
+        } as { [key: string]: string | undefined })[newItem.status];
       }
-      data.find(w => w.title === newItem.type)!.list.push(newItem);
+      data.find((w) => w.title === newItem.type)!.list.push(newItem);
     });
     return data;
   }
 
-  loadData() {
-    if (this.loading) return;
+  loadData(): void {
+    if (this.loading) {
+      return;
+    }
     this.loading = true;
     setTimeout(() => {
+      const now = new Date();
       this.data = this.updateNoticeData([
         {
           id: '000000001',
           avatar: 'https://gw.alipayobjects.com/zos/rmsportal/ThXAXghbEsBCCSDihZxY.png',
           title: '你收到了 14 份新周报',
-          datetime: '2017-08-09',
+          datetime: add(now, { days: 10 }),
           type: '通知',
         },
         {
           id: '000000002',
           avatar: 'https://gw.alipayobjects.com/zos/rmsportal/OKJXDXrmkNshAMvwtvhu.png',
           title: '你推荐的 曲妮妮 已通过第三轮面试',
-          datetime: '2017-08-08',
+          datetime: add(now, { days: -3 }),
           type: '通知',
         },
         {
           id: '000000003',
           avatar: 'https://gw.alipayobjects.com/zos/rmsportal/kISTdvpyTAhtGxpovNWd.png',
           title: '这种模板可以区分多种通知类型',
-          datetime: '2017-08-07',
+          datetime: add(now, { months: -3 }),
           read: true,
           type: '通知',
         },
@@ -100,7 +110,7 @@ export class HeaderNotifyComponent {
           id: '000000004',
           avatar: 'https://gw.alipayobjects.com/zos/rmsportal/GvqBnKhFgObvnSGkDsje.png',
           title: '左侧图标用于区分不同的类型',
-          datetime: '2017-08-07',
+          datetime: add(now, { years: -1 }),
           type: '通知',
         },
         {
@@ -167,16 +177,17 @@ export class HeaderNotifyComponent {
           type: '待办',
         },
       ]);
+
       this.loading = false;
       this.cdr.detectChanges();
-    }, 1000);
+    }, 500);
   }
 
-  clear(type: string) {
+  clear(type: string): void {
     this.msg.success(`清空了 ${type}`);
   }
 
-  select(res: any) {
+  select(res: NoticeIconSelect): void {
     this.msg.success(`点击了 ${res.title} 的 ${res.item.title}`);
   }
 }
