@@ -1,4 +1,4 @@
-import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from '@angular/core';
 
 @Component({
   selector: 'header-task',
@@ -20,13 +20,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
         <div *ngIf="loading" class="mx-lg p-lg"><nz-spin></nz-spin></div>
         <nz-card *ngIf="!loading" nzTitle="Notifications" nzBordered="false" class="ant-card__body-nopadding">
           <ng-template #extra><i nz-icon nzType="plus"></i></ng-template>
-          <div
-            nz-row
-            [nzType]="'flex'"
-            [nzJustify]="'center'"
-            [nzAlign]="'middle'"
-            class="py-sm bg-grey-lighter-h point"
-          >
+          <div nz-row [nzJustify]="'center'" [nzAlign]="'middle'" class="py-sm pr-md point bg-grey-lighter-h">
             <div nz-col [nzSpan]="4" class="text-center">
               <nz-avatar [nzSrc]="'./assets/tmp/img/1.png'"></nz-avatar>
             </div>
@@ -35,13 +29,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
               <p class="mb0">Please tell me what happened in a few words, don't go into details.</p>
             </div>
           </div>
-          <div
-            nz-row
-            [nzType]="'flex'"
-            [nzJustify]="'center'"
-            [nzAlign]="'middle'"
-            class="py-sm bg-grey-lighter-h point"
-          >
+          <div nz-row [nzJustify]="'center'" [nzAlign]="'middle'" class="py-sm pr-md point bg-grey-lighter-h">
             <div nz-col [nzSpan]="4" class="text-center">
               <nz-avatar [nzSrc]="'./assets/tmp/img/2.png'"></nz-avatar>
             </div>
@@ -50,13 +38,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
               <p class="mb0">ハルカソラトキヘダツヒカリ</p>
             </div>
           </div>
-          <div
-            nz-row
-            [nzType]="'flex'"
-            [nzJustify]="'center'"
-            [nzAlign]="'middle'"
-            class="py-sm bg-grey-lighter-h point"
-          >
+          <div nz-row [nzJustify]="'center'" [nzAlign]="'middle'" class="py-sm pr-md point bg-grey-lighter-h">
             <div nz-col [nzSpan]="4" class="text-center">
               <nz-avatar [nzSrc]="'./assets/tmp/img/3.png'"></nz-avatar>
             </div>
@@ -65,13 +47,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
               <p class="mb0">请告诉我，我应该说点什么好？</p>
             </div>
           </div>
-          <div
-            nz-row
-            [nzType]="'flex'"
-            [nzJustify]="'center'"
-            [nzAlign]="'middle'"
-            class="py-sm bg-grey-lighter-h point"
-          >
+          <div nz-row [nzJustify]="'center'" [nzAlign]="'middle'" class="py-sm pr-md point bg-grey-lighter-h">
             <div nz-col [nzSpan]="4" class="text-center">
               <nz-avatar [nzSrc]="'./assets/tmp/img/4.png'"></nz-avatar>
             </div>
@@ -80,13 +56,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
               <p class="mb0">Please tell me what happened in a few words, don't go into details.</p>
             </div>
           </div>
-          <div
-            nz-row
-            [nzType]="'flex'"
-            [nzJustify]="'center'"
-            [nzAlign]="'middle'"
-            class="py-sm bg-grey-lighter-h point"
-          >
+          <div nz-row [nzJustify]="'center'" [nzAlign]="'middle'" class="py-sm pr-md point bg-grey-lighter-h">
             <div nz-col [nzSpan]="4" class="text-center">
               <nz-avatar [nzSrc]="'./assets/tmp/img/5.png'"></nz-avatar>
             </div>
@@ -96,9 +66,7 @@ import { Component, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/
             </div>
           </div>
           <div nz-row>
-            <div nz-col [nzSpan]="24" class="pt-md border-top-1 text-center text-grey point">
-              See All
-            </div>
+            <div nz-col [nzSpan]="24" class="pt-md border-top-1 text-center text-grey point">See All</div>
           </div>
         </nz-card>
       </div>
@@ -111,7 +79,7 @@ export class HeaderTaskComponent {
 
   constructor(private cdr: ChangeDetectorRef) {}
 
-  change() {
+  change(): void {
     setTimeout(() => {
       this.loading = false;
       this.cdr.detectChanges();
