@@ -1,11 +1,12 @@
-import { Component } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
 
 @Component({
   selector: 'layout-passport',
   templateUrl: './passport.component.html',
   styleUrls: ['./passport.component.less'],
 })
-export class LayoutPassportComponent {
+export class LayoutPassportComponent implements OnInit {
   links = [
     {
       title: '帮助',
@@ -20,4 +21,10 @@ export class LayoutPassportComponent {
       href: '',
     },
   ];
+
+  constructor(@Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService) {}
+
+  ngOnInit(): void {
+    this.tokenService.clear();
+  }
 }
