@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '@shared';
 
+import { ExceptionModule as DelonExceptionModule } from '@delon/abc/exception';
 import { ExceptionRoutingModule } from './exception-routing.module';
 
 import { Exception403Component } from './403.component';
@@ -12,8 +13,9 @@ const COMPONENTS = [Exception403Component, Exception404Component, Exception500Co
 const COMPONENTS_NOROUNT = [];
 
 @NgModule({
-  imports: [SharedModule, ExceptionRoutingModule],
+  imports: [SharedModule, DelonExceptionModule, ExceptionRoutingModule],
   declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
   entryComponents: COMPONENTS_NOROUNT,
 })
-export class ExceptionModule {}
+export class ExceptionModule {
+}
