@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { SimpleGuard } from '@delon/auth';
 import { environment } from '@env/environment';
 // layout
-import { LayoutDefaultComponent } from '../layout/default/default.component';
 import { LayoutPassportComponent } from '../layout/passport/passport.component';
 // dashboard pages
 import { DashboardComponent } from './dashboard/dashboard.component';
@@ -12,11 +10,13 @@ import { UserLoginComponent } from './passport/login/login.component';
 // single pages
 import { CallbackComponent } from './callback/callback.component';
 import { UserLockComponent } from './passport/lock/lock.component';
+import { SimpleGuard } from "@delon/auth";
+import { LayoutBasicComponent } from "../layout/basic/basic.component";
 
 const routes: Routes = [
   {
     path: '',
-    component: LayoutDefaultComponent,
+    component: LayoutBasicComponent,
     canActivate: [SimpleGuard],
     children: [
       {path: '', redirectTo: 'dashboard', pathMatch: 'full'},
@@ -52,12 +52,11 @@ const routes: Routes = [
   imports: [
     RouterModule.forRoot(
       routes, {
-    useHash: environment.useHash,
-    // NOTICE: If you use `reuse-tab` component and turn on keepingScroll you can set to `disabled`
-    // Pls refer to https://ng-alain.com/components/reuse-tab
-    scrollPositionRestoration: 'top',
-    relativeLinkResolution: 'legacy'
-}
+        useHash: environment.useHash,
+        // NOTICE: If you use `reuse-tab` component and turn on keepingScroll you can set to `disabled`
+        // Pls refer to https://ng-alain.com/components/reuse-tab
+        scrollPositionRestoration: 'top'
+      }
     )],
   exports: [RouterModule],
 })
