@@ -7,33 +7,23 @@ import { environment } from '@env/environment';
 import { preloaderFinished } from '@delon/theme';
 preloaderFinished();
 
-import { hmrBootstrap } from './hmr';
+import { NzSafeAny } from "ng-zorro-antd/core/types";
 
 if (environment.production) {
   enableProdMode();
 }
 
-const bootstrap = () => {
-  return platformBrowserDynamic()
-    .bootstrapModule(AppModule, {
-      defaultEncapsulation: ViewEncapsulation.Emulated,
-    })
-    .then(res => {
-      if ((window as any).appBootstrap) {
-        (window as any).appBootstrap();
-      }
-      return res;
-    });
-};
 
-if (environment.hmr) {
-  // tslint:disable-next-line: no-string-literal
-  if (module['hot']) {
-    hmrBootstrap(module, bootstrap);
-  } else {
-    console.error('HMR is not enabled for webpack-dev-server!');
-    console.log('Are you using the --hmr flag for ng serve?');
-  }
-} else {
-  bootstrap();
-}
+platformBrowserDynamic()
+  .bootstrapModule(AppModule, {
+    defaultEncapsulation: ViewEncapsulation.Emulated,
+    preserveWhitespaces: false,
+  })
+  .then((res) => {
+    const win = window as NzSafeAny;
+    if (win && win.appBootstrap) {
+      win.appBootstrap();
+    }
+    return res;
+  })
+  .catch((err) => console.error(err));
